use size_t instead of int for array size

+++ src/t8_schemes/t8_default/t8_default_hex/t8_default_hex_cxx.hxx
@@ -581,7 +581,7 @@ public:
    */
   virtual void        t8_element_reference_coords (const t8_element_t *elem,
                                                    const double *ref_coords,
-                                                   const int num_coords,
+                                                   const size_t num_coords,
                                                    double *out_coords)
     const;
